+++ MoreGrassStarters/docs/release-notes.md
@@ -1,6 +1,12 @@
 ﻿﻿[← back to readme](README.md)
 
 # Release notes
+## 1.1.0
+Released 23 December 2021 for SMAPI 3.13.0 or later. Updated by Pathoschild.
+
+* Updated for Stardew Valley 1.5.5.
+* Added support for custom grass spread.
+
 ## 1.0.6
 Released 11 September 2021 for SMAPI 3.12.5 or later. Updated by Pathoschild.